backend/requirements/dev.txt
-r base.txt

django-extensions==3.2.3
pytest==8.1.1
pytest-django==4.8.0
model-bakery==1.18.0
coverage==7.4.4
